hdsutils: force get_hds_values to return strings instead of bytes
This should really be cleverer than this -- I'm not sure its 100%
python2 safe, and it should alsmost certainly be using a specific
encoding (instead of the default), and have some error handling
setup. This is enough to get started with though I think.

Currently, I think everyone will get unicode strings for the output.