wrapper: delete $ADAMDIR/<commandname>.sdf after running command.
Starlink commands can produce different output parameters for
different input options. E.g, if you run KAPPA's stats with
order=True, it will create a 'median' value in the output
file. However, if you then run stats with order=False it will not
blank out or overwrite this value in the $ADAMDIR/stats.sdf. For this
package, which just reads everything from the output file in the ADAM
dir and returns it to the user as tuple, this behaviour is very
misleading. I've checked with David and Malcolm, and its definitely
not considered a bug for Starlink, so I'm working around it here by
just deleting the whole file every time.

This seems like a very inefficient solution, but I'm not sure what
else I can do: RESET doesn't reset output parameters so that can't be
used. This does have the happy effect of removing a lot of the 'use
your last used value as the default' behaviour which can be very
confusing in scripts.

The file is also deleted when there's an error, as I've found that
it causes really weird problems sometimes when that happens.