Decoding HDS strings: handle errors on decoding.
I have mangaged to find at least one occassion where the bytestring
returned can't be cleanly decoded, so for now use the errors='replace'
option to <str>.decode. This will replace the character that can't be
decoded with the appropriate unknown character.

(I think the bytestring in question is a bug, where minwcs in KAPPA stats
is given a silly value when you run stats on a non-existent plane
by typoing an NDF section, but this wrapper still needs to handle it cleanly.)